# Improve memory scanning performance a bit