--- a/src/gc.cpp
+++ b/src/gc.cpp
@@ -203,8 +203,11 @@ HeapBlock *Allocator::add_heap_block() {
     m_blocks.set(block);
     add_free_block(block);
     m_free_cells += cell_count_per_block();
-    Heap::the().m_free_cells += cell_count_per_block();
-    Heap::the().m_total_cells += cell_count_per_block();
+    auto &heap = Heap::the();
+    heap.m_free_cells += cell_count_per_block();
+    heap.m_total_cells += cell_count_per_block();
+    heap.m_lowest_pointer_address = std::min(heap.m_lowest_pointer_address, reinterpret_cast<uintptr_t>(block));
+    heap.m_highest_pointer_address = std::max(heap.m_highest_pointer_address, reinterpret_cast<uintptr_t>(block) + HEAP_BLOCK_SIZE);
     return block;
 }
 
@@ -245,6 +248,8 @@ NO_SANITIZE_ADDRESS void Heap::scan_memory(Cell::Visitor &visitor, void *start, 
             continue;
         if ((reinterpret_cast<uintptr_t>(potential_cell) & 0b111) != 0b000)
             continue;
+        if (reinterpret_cast<uintptr_t>(potential_cell) < m_lowest_pointer_address || reinterpret_cast<uintptr_t>(potential_cell) > m_highest_pointer_address)
+            continue;
         if (is_a_heap_cell_in_use(potential_cell))
             visitor.visit(potential_cell);
     }
@@ -256,6 +261,8 @@ NO_SANITIZE_ADDRESS void Heap::scan_memory(Cell::Visitor &visitor, void *start, 
         if (!potential_cell)
             continue;
         if ((reinterpret_cast<uintptr_t>(potential_cell) & 0b111) != 0b000)
+            continue;
+        if (reinterpret_cast<uintptr_t>(potential_cell) < m_lowest_pointer_address || reinterpret_cast<uintptr_t>(potential_cell) > m_highest_pointer_address)
             continue;
         if (is_a_heap_cell_in_use(potential_cell))
             visitor.visit(potential_cell);