This shows small improvement when running a program with lots of GC runs, including the self-hosted compiler.

```
→ hyperfine --warmup 10 'bin/nat_master -c h1 examples/hello.rb' 'bin/nat -c h2 examples/hello.rb'

Benchmark 1: bin/nat_master -c h1 examples/hello.rb
  Time (mean ± σ):     927.4 ms ±  32.2 ms    [User: 720.6 ms, System: 156.1 ms]
  Range (min … max):   881.7 ms … 991.4 ms    10 runs

Benchmark 2: bin/nat -c h2 examples/hello.rb
  Time (mean ± σ):     878.5 ms ±   5.2 ms    [User: 723.1 ms, System: 154.5 ms]
  Range (min … max):   870.5 ms … 884.7 ms    10 runs

Summary
  bin/nat -c h2 examples/hello.rb ran
    1.06 ± 0.04 times faster than bin/nat_master -c h1 examples/hello.rb
```